% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_columns.R
\name{check_columns}
\alias{check_columns}
\title{check_columns() Function}
\usage{
check_columns(data, var)
}
\arguments{
\item{data}{A data frame containing measurement values and dates. Column names
should be formatted as "ID", "Measurement name 1", "Date Measurement name 1",
..., "Measurement name k", "Date Measurement name k".}

\item{var}{A character string specifying the type of measurement to be processed
\itemize{
\item either "Body Weight" or "Outcome".
}}
}
\value{
This function returns a list:
\item{data}{cleaned dataset}
\item{mismatches}{any incorrect column names that were found}
}
\description{
This function makes sure that the column names are appropriately named. If any
incorrect column names are found, it asks user to change it.
}
