% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIDs.R
\name{checkIDs}
\alias{checkIDs}
\title{checkIDs() Function}
\usage{
checkIDs(data, birthdata, var)
}
\arguments{
\item{data}{A data frame containing measurement values and dates. Column names
should be formatted as "ID", "Measurement name 1", "Date Measurement name 1",
..., "Measurement name k", "Date Measurement name k".}

\item{birthdata}{Contains cleaned birth data.}

\item{var}{A character string specifying the type of measurement to be processed
\itemize{
\item either "Body Weight" or "Outcome".
}}
}
\value{
A list containing:
\item{invalid_ids}{table of invalid IDs found}
\item{remaining_valid_ids}{any remaining valid IDs not found in dataset}
\item{changes_made}{number of invalid IDs found and changed}
}
\description{
This function makes sure that the IDs in a dataset align with those found in
the cleaned birth data set.
}
