% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_birth_data.R
\name{clean_birth_data}
\alias{clean_birth_data}
\title{clean_birth_data() Function}
\usage{
clean_birth_data(data)
}
\arguments{
\item{data}{Unclean dataset}
}
\value{
Returns a clean, pre-processed dataset
}
\description{
clean_birth_data() Function
}
