% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_mouse_data.R
\name{clean_mouse_data}
\alias{clean_mouse_data}
\title{clean_mouse_data() Function}
\usage{
clean_mouse_data(file_path)
}
\arguments{
\item{file_path}{Path of the excel file that contains your data. The file
must contain three tabs - birth data, body weight data, and outcome data.}
}
\value{
A list containing:
\item{birth}{cleaned birth data}
\item{weight}{cleaned weight data}
\item{outcome}{cleaned outcome data}
}
\description{
This function cleans the mouse data by obtaining sheet information on the excel
file. It then uses this sheet information and the loaded excel file to
call the individual clean functions to fully clean it.
}
