% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_outcome_data.R
\name{clean_outcome_data}
\alias{clean_outcome_data}
\title{clean_outcome_data() Function}
\usage{
clean_outcome_data(data, birthdata)
}
\arguments{
\item{data}{Weight data to be cleaned.}

\item{birthdata}{Contains cleaned birth data.}
}
\value{
A cleaned outcome data set with correct column names, IDs,
body weight value types, and date format
}
\description{
clean_outcome_data() Function
}
