% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_weight_data.R
\name{clean_weight_data}
\alias{clean_weight_data}
\title{clean_weight_data() Function}
\usage{
clean_weight_data(data, birthdata)
}
\arguments{
\item{data}{A data frame containing weight data values and dates. Column names
should be formatted as "ID", "Body Weight 1", "Date Body Weight 1",
..., "Body Weight k", "Date Body Weight k".}

\item{birthdata}{Contains cleaned birth data.}
}
\value{
A cleaned weight data set with correct column names, IDs,
body weight value types, and date format
}
\description{
This function cleans weight data by checking column names, IDs, and
data types. If any errors are found, the function prompts the user
to fix it.
}
