% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_treatment_info.R
\name{create_treatment_info}
\alias{create_treatment_info}
\title{create_treatment_info() Function}
\usage{
create_treatment_info(birth_data)
}
\arguments{
\item{birth_data}{This is the cleaned birth data set from the clean_mouse_data
function output}
}
\value{
This function takes cleaned birth data as an input and outputs
a dataset with only ID and treatment assignment.
}
\description{
create_treatment_info() Function
}
