% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelsheetinfo.R
\name{excelsheetinfo}
\alias{excelsheetinfo}
\title{excelsheetinfo() Function}
\usage{
excelsheetinfo(path)
}
\arguments{
\item{path}{Path of the excel sheet.}
}
\value{
This function returns a mapping of each sheet in the excel to
a particular data set type (Birth, Body Weight, or Outcome).
}
\description{
excelsheetinfo() Function
}
