% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyoutliers.R
\name{identifyoutliers}
\alias{identifyoutliers}
\title{identifyoutliers() Function}
\usage{
identifyoutliers(data, var, treatment_info)
}
\arguments{
\item{data}{Data to be analyzed}

\item{var}{Type of measurement - "Body Weight" or "Outcome"}

\item{treatment_info}{Data set that maps ID to treatment assignment. This can
be made by inputting birthdata into create_treatment_info() function.}
}
\value{
A list containing:
\item{model}{Fitted time-series model}
\item{outliers}{Data frame of identified outliers}
\item{plot}{ggplot plot of the data with outliers highlighted}
}
\description{
identifyoutliers() Function
}
