% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightloss_check.R
\name{weightloss_check}
\alias{weightloss_check}
\title{weightloss_check() Function}
\usage{
weightloss_check(data, var, threshold = 20)
}
\arguments{
\item{data}{Data to be analyzed}

\item{var}{Type of measurement - "Body Weight" or "Outcome"}

\item{threshold}{Percentage weight loss threshold for removal (default = 20\%)}
}
\value{
A list containing:
\item{sac_mice}{Summary table of all mice that exceed the weight threshold}
\item{all_mice}{Summary table of all mice and weight}
}
\description{
weightloss_check() Function
}
