% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_columns.R
\name{check_columns}
\alias{check_columns}
\title{check_columns() Function}
\usage{
check_columns(data, var)
}
\arguments{
\item{data}{Data set to be checked}

\item{var}{Type of data set that is being checked - either "Body Weight" or "Outcome"}
}
\value{
The purpose of this function is to check that the Body Weight or
Outcome column names are as expected before we proceed with analysis.
}
\description{
check_columns() Function
}
