% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIDs.R
\name{checkIDs}
\alias{checkIDs}
\title{checkIDs() Function}
\usage{
checkIDs(data, birthdata, var)
}
\arguments{
\item{data}{The current dataset being checked}

\item{birthdata}{Reference data set containing the IDs}

\item{var}{Type of measurement - "Body Weight" or "Outcome"}
}
\value{
A list containing the modified data and summary tables
}
\description{
checkIDs() Function
}
