% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkquantvars.R
\name{checkquantvars}
\alias{checkquantvars}
\title{checkquantvars() Function}
\usage{
checkquantvars(data, name)
}
\arguments{
\item{data}{Data set to be checked}

\item{name}{Common name of columns that should contain quantitative variables.
Should be "Outcome" or "Body Weight"}
}
\value{
This function takes in a data set and makes sure that all quantitative
variable columns contain numeric values. If it finds a non-numeric value,
it asks user to input NA or a numeric value.
}
\description{
checkquantvars() Function
}
