% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_birth_data.R
\name{clean_birth_data}
\alias{clean_birth_data}
\title{clean_birth_data() Function}
\usage{
clean_birth_data(data)
}
\arguments{
\item{data}{A data frame containing birth information. Column names
should be formatted as "ID", "Sex", "Num", "Treatment"}
}
\value{
Returns a cleaned birth dataset.
}
\description{
This function cleans birth data by checking treatment assignments and IDs.
}
