% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_outcome_data.R
\name{clean_outcome_data}
\alias{clean_outcome_data}
\title{clean_outcome_data() Function}
\usage{
clean_outcome_data(data, birthdata)
}
\arguments{
\item{data}{A data frame containing outcome values and dates. Column names
should be formatted as "ID", "Outcome 1", "Date Outcome 1",
..., "Outcome k", "Date Outcome k".}

\item{birthdata}{Contains cleaned birth data.}
}
\value{
A cleaned outcome data set with correct column names and IDs
}
\description{
This function cleans outcome data by checking column names, IDs, and
data types. If any errors are found, the function prompts the user
to fix it.
}
