% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelsheetinfo.R
\name{excelsheetinfo}
\alias{excelsheetinfo}
\title{excelsheetinfo() Function}
\usage{
excelsheetinfo(path)
}
\arguments{
\item{path}{The path of the excel file to be analyzed.}
}
\value{
This function returns a mapping of each sheet in the excel to
a particular data set type (Birth, Body Weight, or Outcome).
}
\description{
This function reads in the excel file at the given path and then prompts the
user to identify which sheets inside the excel file are which data set type.
(Birth, Body Weight, or Outcome)
}
