% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyoutliers.R
\name{identifyoutliers}
\alias{identifyoutliers}
\title{identifyoutliers() Function}
\usage{
identifyoutliers(data, var, treatment_info)
}
\arguments{
\item{data}{A data frame containing measurement values and dates. Column names
should be formatted as "ID", "Measurement name 1", "Date Measurement name 1",
..., "Measurement name k", "Date Measurement name k".}

\item{var}{A character string specifying the type of measurement to be processed
\itemize{
\item either "Body Weight" or "Outcome".
}}

\item{treatment_info}{Data set that maps ID to treatment assignment. This can
be made by inputting birthdata into create_treatment_info() function.}
}
\value{
A list containing:
\item{model}{Fitted time-series model}
\item{outliers}{Data frame of identified outliers}
\item{plot}{ggplot plot of the data with outliers highlighted}
}
\description{
This function builds a time series model based off of inputted data and using
standardized residuals, it identifies potential outlier data.
}
