% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_indmeasurement.R
\name{plot_indmeasurement}
\alias{plot_indmeasurement}
\title{plot_indmeasurement() Function}
\usage{
plot_indmeasurement(data, var, treatment_info)
}
\arguments{
\item{data}{Data to be analyzed}

\item{var}{Type of measurement - "Body Weight" or "Outcome"}

\item{treatment_info}{Data set that maps ID to treatment assignment. This can
be made by inputting birthdata into create_treatment_info() function.}
}
\value{
A list containing:
\item{p_tmt}{ggplot2 visualization of treatment group}
\item{p_plac}{ggplot2 visualization of placebo group}
}
\description{
plot_indmeasurement() Function
}
