% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessdata.R
\name{preprocessdata}
\alias{preprocessdata}
\title{preprocessdata() Function}
\usage{
preprocessdata(data, var)
}
\arguments{
\item{data}{Data set to be preprocessed.}

\item{var}{Type of measurement - "Body Weight" or "Outcome"}
}
\value{
Modified data set proper for analysis.
}
\description{
preprocessdata() Function
}
